add game 2 mode

+++ game/models.py
@@ -12,7 +12,7 @@ import random
 from faker import Faker
 from customuser.models import User
 import puzzle_game.settings
-
+from django_random_queryset import RandomManager
 from django.core.files.base import ContentFile
 from django.core.files.storage import default_storage
 
@@ -39,12 +39,17 @@ class Image(models.Model):
     def __str__(self):
         return '{id}: {title}'.format(id=self.id, title=self.title)
 
+class GameImage(models.Model):
+    image = models.ImageField(blank=True, upload_to='game_images/')
+    objects = RandomManager()
+
 
 class Game(models.Model):
     player = models.ForeignKey(User, blank=True, on_delete=models.CASCADE, null=True)
     # lev = models.ForeignKey(Level, on_delete=models.PROTECT, null=True, blank=True)
     # image = models.ForeignKey(Image, on_delete=models.PROTECT)
     level = models.IntegerField(default=0)
+    game_type = models.IntegerField(blank=True,null=True)
     image = models.ImageField(blank=True, upload_to='games/')
     date = models.DateTimeField(auto_now_add=True, blank=True)
     start = models.DateTimeField(auto_now_add=True, blank=True)
@@ -65,17 +70,42 @@ class Game(models.Model):
         super().save(*args, **kwargs)
 
     def create_qr(self):
+        image = None
         name = str(uuid.uuid4())
-        color = tuple(random.choice(range(100)) for _ in range(3))
-        fake = Faker()
-        print(settings.STATIC_ROOT)
-        qr = segno.make_qr(fake.text(2000), version=33, error='L', mask=3)
-        qr.save(f'{settings.MEDIA_ROOT}\games\{name}.png', scale=10, dark=color, light=(240,240,240), border=0)
-        image_bytes_stream = BytesIO()
-        temp_image = PILImage.open(f'{settings.MEDIA_ROOT}\games\{name}.png')
-        temp_image.seek(0)
-        temp_image = temp_image.resize((800, 800))
-        temp_image.save(image_bytes_stream, format='png', quality=100)
-        default_storage.delete('games/{}.png'.format(name))
-        image = File(image_bytes_stream, name='{}.png'.format(name))
+        if self.game_type == 2:
+            color = tuple(random.choice(range(100)) for _ in range(3))
+            fake = Faker()
+            qr = segno.make_qr(fake.text(2000), version=33, error='L', mask=3)
+            qr.save(f'{settings.MEDIA_ROOT}\games\{name}.png', scale=10, dark=color, light=(240, 240, 240), border=0)
+            image_bytes_stream = BytesIO()
+            temp_image = PILImage.open(f'{settings.MEDIA_ROOT}\games\{name}.png')
+            temp_image.seek(0)
+            temp_image = temp_image.resize((800, 800))
+            temp_image.save(image_bytes_stream, format='png', quality=100)
+            default_storage.delete('games/{}.png'.format(name))
+            image = File(image_bytes_stream, name='{}.png'.format(name))
+        elif self.game_type == 1:
+            if self.player:
+                print(self.player)
+                images = GameImage.objects.all()
+                for img in images:
+                    if self.player.last_image != img.id :
+                        temp_image = img
+                        self.player.last_image = img.id
+                        self.player.save()
+                        break
+                    else:
+                        temp_image = GameImage.objects.random(1).first()
+                        self.player.last_image = temp_image.id
+                        self.player.save()
+            else:
+                temp_image = GameImage.objects.random(1).first()
+
+            image_bytes_stream = BytesIO()
+            temp_image = PILImage.open(f'{settings.BASE_DIR}{temp_image.image.url}')
+            temp_image.seek(0)
+            # temp_image = temp_image.resize((800, 800))
+            temp_image.save(image_bytes_stream, format='png', quality=100)
+            image = File(image_bytes_stream, name='{}.png'.format(name))
+
         return image
